Refact: moved function and cleaned up code.
