+++ osl_dynamics/data/__init__.py
@@ -10,6 +10,7 @@ New users may find the following tutorials helpful:
   /tutorials_build/data_preparation.html>`_
 """
 
-from osl_dynamics.data.base import Data, SessionLabels, load_tfrecord_dataset
+from osl_dynamics.data.base import Data, SessionLabels
+from osl_dynamics.data.rw import load_tfrecord_dataset
 
-__all__ = ["Data"]
+__all__ = ["Data", "SessionLabels", "load_tfrecord_dataset"]
