+++ osl_dynamics/data/rw.py
@@ -10,6 +10,8 @@ import mat73
 import numpy as np
 import scipy.io
 
+from osl_dynamics.utils import misc
+
 _logger = logging.getLogger("osl-dynamics")
 _allowed_ext = [".npy", ".mat", ".txt", ".fif"]
 
@@ -301,3 +303,202 @@ def loadmat(filename, return_dict=False):
             mat = mat[fields[0]]
 
     return mat
+
+
+def load_tfrecord_dataset(
+    tfrecord_dir,
+    batch_size,
+    shuffle=True,
+    validation_split=None,
+    concatenate=True,
+    drop_last_batch=False,
+    buffer_size=100000,
+    keep=None,
+):
+    """Load a TFRecord dataset.
+
+    Parameters
+    ----------
+    tfrecord_dir : str
+        Directory containing the TFRecord datasets.
+    batch_size : int
+        Number sequences in each mini-batch which is used to train the model.
+    shuffle : bool, optional
+        Should we shuffle sequences (within a batch) and batches.
+    validation_split : float, optional
+        Ratio to split the dataset into a training and validation set.
+    concatenate : bool, optional
+        Should we concatenate the datasets for each array?
+    drop_last_batch : bool, optional
+        Should we drop the last batch if it is smaller than the batch size?
+    buffer_size : int, optional
+        Buffer size for shuffling a TensorFlow Dataset. Smaller values will lead
+        to less random shuffling but will be quicker. Default is 100000.
+    keep : list of int, optional
+        List of session indices to keep. If :code:`None`, then all sessions
+        are kept.
+
+    Returns
+    -------
+    dataset : tf.data.Dataset or tuple
+        Dataset for training or evaluating the model along with the validation
+        set if :code:`validation_split` was passed.
+    """
+    import tensorflow as tf  # moved here to avoid slow imports
+
+    tf_record_config = misc.load(f"{tfrecord_dir}/tfrecord_config.pkl")
+    identifier = tf_record_config["identifier"]
+    sequence_length = tf_record_config["sequence_length"]
+    n_channels = tf_record_config["n_channels"]
+    session_labels = tf_record_config["session_labels"]
+    n_sessions = tf_record_config["n_sessions"]
+
+    keep = keep or list(range(n_sessions))
+
+    # Helper function for parsing training examples
+    def _parse_example(example):
+        feature_names = ["data"]
+        tensor_shapes = {
+            "data": [sequence_length, n_channels],
+        }
+        # Add session labels if there are any
+        for feature_name in session_labels:
+            feature_names.append(feature_name)
+            tensor_shapes[feature_name] = [sequence_length]
+
+        feature_description = {
+            name: tf.io.FixedLenFeature([], tf.string) for name in feature_names
+        }
+        parsed_example = tf.io.parse_single_example(
+            example,
+            feature_description,
+        )
+        return {
+            name: tf.ensure_shape(
+                tf.io.parse_tensor(tensor, tf.float32), tensor_shapes[name]
+            )
+            for name, tensor in parsed_example.items()
+        }
+
+    tfrecord_paths = (
+        f"{tfrecord_dir}"
+        "/dataset_{array:0{v}d}-of-{n_session:0{v}d}"
+        f".{identifier}.tfrecord"
+    )
+    tfrecord_filenames = []
+    for i in keep:
+        filepath = tfrecord_paths.format(
+            array=i,
+            n_session=n_sessions - 1,
+            v=len(str(n_sessions - 1)),
+        )
+        tfrecord_filenames.append(filepath)
+
+    # Create the TFRecord dataset
+    if concatenate:
+        tfrecord_filenames = tf.data.Dataset.from_tensor_slices(tfrecord_filenames)
+
+        if shuffle:
+            # First shuffle the shards
+            tfrecord_filenames = tfrecord_filenames.shuffle(len(tfrecord_filenames))
+
+            # Create the TFRecord dataset
+            full_dataset = tfrecord_filenames.interleave(
+                tf.data.TFRecordDataset, num_parallel_calls=tf.data.AUTOTUNE
+            )
+
+            # Parse the examples
+            full_dataset = full_dataset.map(_parse_example)
+
+            # Shuffle sequences
+            full_dataset = full_dataset.shuffle(buffer_size)
+
+            # Group into batches
+            full_dataset = full_dataset.batch(
+                batch_size, drop_remainder=drop_last_batch
+            )
+
+            # Shuffle batches
+            full_dataset = full_dataset.shuffle(buffer_size)
+
+        else:
+            # Create the TFRecord dataset
+            full_dataset = tfrecord_filenames.interleave(
+                tf.data.TFRecordDataset, num_parallel_calls=tf.data.AUTOTUNE
+            )
+
+            # Parse the examples
+            full_dataset = full_dataset.map(_parse_example)
+
+            # Group into batches
+            full_dataset = full_dataset.batch(
+                batch_size, drop_remainder=drop_last_batch
+            )
+
+        if validation_split is None:
+            # Return the dataset
+            return full_dataset.prefetch(tf.data.AUTOTUNE)
+
+        else:
+            # Calculate how many batches should be in the training dataset
+            dataset_size = dtf.get_n_batches(full_dataset)
+            training_dataset_size = round((1.0 - validation_split) * dataset_size)
+
+            # Split the dataset into training and validation datasets
+            training_dataset = full_dataset.take(training_dataset_size)
+            validation_dataset = full_dataset.skip(training_dataset_size)
+            _logger.info(
+                f"{training_dataset_size} batches in training dataset, "
+                + f"{dataset_size - training_dataset_size} batches in the validation "
+                + "dataset."
+            )
+            return training_dataset.prefetch(
+                tf.data.AUTOTUNE
+            ), validation_dataset.prefetch(tf.data.AUTOTUNE)
+
+    # Otherwise create a dataset for each array separately
+    else:
+        full_datasets = []
+        for filename in tfrecord_filenames:
+            ds = tf.data.TFRecordDataset(filename)
+
+            # Parse the examples
+            ds = ds.map(_parse_example)
+
+            if shuffle:
+                # Shuffle sequences
+                ds = ds.shuffle(buffer_size)
+
+            # Group into batches
+            ds = ds.batch(batch_size, drop_remainder=drop_last_batch)
+
+            if shuffle:
+                # Shuffle batches
+                ds = ds.shuffle(buffer_size)
+
+            full_datasets.append(ds.prefetch(tf.data.AUTOTUNE))
+
+        if validation_split is None:
+            # Return the full dataset for each array
+            return full_datasets
+
+        else:
+            # Split the dataset for each array separately
+            training_datasets = []
+            validation_datasets = []
+            for i, ds in enumerate(full_datasets):
+                # Calculate how many batches should be in the training dataset
+                dataset_size = dtf.get_n_batches(ds)
+                training_dataset_size = round((1.0 - validation_split) * dataset_size)
+
+                # Split the dataset into training and validation datasets
+                training_datasets.append(ds.take(training_dataset_size))
+                validation_datasets.append(ds.skip(training_dataset_size))
+                _logger.info(
+                    f"Session {i}: "
+                    + f"{training_dataset_size} batches in training dataset, "
+                    + f"{dataset_size - training_dataset_size} batches in the validation "
+                    + "dataset."
+                )
+
+            return training_datasets, validation_datasets
